refactor: HttpQuery 와 Query 분
