+++ tomcat/src/test/java/org/apache/coyote/http11/HttpQueryTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.techcourse.exception.client.BadRequestException;
+import org.apache.coyote.http11.query.HttpQuery;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
